openapi/requirements.txt
openai==1.6.0
weave==0.52.8
python-dotenv==1.0.0
pytest==7.4.3
